phase1/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
aiofiles==24.1.0
mistune==3.0.2
difflib2==0.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0        # SQLite async driver for tests
greenlet==3.1.1           # required by SQLAlchemy async


